Align aws profile parsing with other SDKs
multiline properties are still not supported though.
