+++ aws-cpp-sdk-core/include/aws/core/auth/signer/AWSAuthSignerBase.h
